requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pdf2docx==0.5.6

# AI and NLP
openai==1.3.7
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Export and Formatting
reportlab==4.0.7
weasyprint==60.2
fpdf2==2.7.6
markdown==3.5.1

# OAuth and Authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlite3

# Utilities
pillow==10.1.0
requests==2.31.0
httpx==0.25.2
validators==0.22.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Security and Rate Limiting
slowapi==0.1.9
cryptography==41.0.7
